requirements.txt
flask
python-dotenv
openai
twilio
requests
pdfminer.six
python-docx
Pillow
